convert LinkInversion to enum class

+++ src/effects/effectparameterslot.cpp
@@ -86,7 +86,8 @@ void EffectParameterSlot::loadEffect(EffectPointer pEffect) {
             m_pControlLoaded->setAndConfirm(1.0);
 
             m_pControlLinkType->set(m_pEffectParameter->getDefaultLinkType());
-            m_pControlLinkInverse->set(m_pEffectParameter->getDefaultLinkInversion());
+            m_pControlLinkInverse->set(
+                static_cast<double>(m_pEffectParameter->getDefaultLinkInversion()));
 
             connect(m_pEffectParameter, SIGNAL(valueChanged(double)),
                     this, SLOT(slotParameterValueChanged(double)));
